fix: make mimetype to file type function standalone
